Add gitattributes w/ export-ignore
This will allow for a production ready version of scrollReveal when archived/exported for release(s) and still allowing a development version within the repository.

This would be very useful because releases are intended for production use and when users are using a packaged release (e.g. as a dependancy), they most likely are not interested in downloading the source code as a whole into production, thus saving server disk space and meeting security protocol. Think of it as a .gitignore for releases.

  - e.g. /test folders, or your bower.json file, or other build code etc., will not be included in the release archieves (zip(s))

See: [git-scm.com - Exporting-Your-Repository](http://git-scm.com/book/en/v2/Customizing-Git-Git-Attributes#Exporting-Your-Repository)