replace karma browserify plugin with rollup
The goal is to be testing the same code that's shipping to users,
since the Browserify code is horribly bloated -- I measured the
file size to be DOUBLE that of rollup. Not to mention that
Browserify was using Babel, whereas the distribution was using
Buble...

Suffice to say, the frustration was worth the confidence that what
is being shipped, is what is being tested. (It also means my code
coverage looks a lot more like the source.)

An example benefit of this strategy, is that I've identified the
Object.assign method as needing a polyfill, and that the spread
operator on NodeLists does not work with Buble.

Some tests are failing in this build, but that will be rectified
shortly, as the next few commits will be focused on a polyfill
for Object.assign....