extract transform matrix values into array
This is getting into a bit of deep shit, in order to finally
resolve #251 and preserve CSS transforms. I've successfully
captured the computed transformation matrix as a 16 value array...
even when the shorthand is provided.

Definitely looking promising, but this is just the first step.
Next I will need to generate a matrix from the `reveal()` config,
which appears to be the biggest hurdle.