replace explicit animate call from initialize
The handler function calls animate in the next animation frame,
so it's unecessary to import and invoke animate explicitly.
Moreover, the call to animate doesn't accurately describe the
flow of logic...

As the handler functions grows beyond simply calling animate
inside request animation frame, the new functionality will still
almost certainly be relevant after intialization; the true need
is not a one-time animate call, but a one-time handler call to
capture the initial state.