refactor dom node utilities error handling
In preparation for a `debug` flag to control whether or not errors
are surfaced, the `logger` utility has been removed from the dom
node utility functions. (loosened coupling)

This effectively means `this` does not need to be passed around as
much, since only the `logger` will need a `this` reference.